+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 Features:
 
 - Added alert on Periwinkle Dye and Bone Dye drop, also added them to Fishing profit tracker when it's obtained while fishing.
+- Track Ice Essence fished up from water on Jerry's Workshop.
 
 ## v1.22.0